+++ tests/_test.py
@@ -1506,3 +1506,20 @@ def test_git_commit_nested_with_other_dotfiles(
     path.write_text(changed[1], encoding="utf-8")
     result = cli((d.main, [COMMIT, path]))
     assert NOTHING_TO_COMMIT not in result.stdout
+
+
+def test_commit_message_path(
+    cli: FixtureCli, make_tree: FixtureMakeTree
+) -> None:
+    """Test correct commit message added on commit.
+
+    :param cli: Cli runner for testing.
+    :param make_tree: Make file tree.
+    """
+    make_tree({P1.dst: {P2.src: P2.contents}})
+    path = P1.dst / P2.src
+    cli((d.main, [ADD, path.parent]), (d.main, [ADD, path]))
+    assert (
+        str(path)
+        in git.Repo(DOTFILES).git.log("HEAD", format="%B", max_count=1).strip()
+    )